backend/requirements.txt
livekit-agents==1.5.2
livekit-plugins-google==1.5.2
firebase-admin
fastapi
uvicorn
python-dotenv
